move files around, fix logging
