+++ src/abacus-ts/lib/signal.ts
