+++ src/abacus-ts/logs.ts
@@ -1,0 +1,5 @@
+import { log } from '@/lib/telemetry';
+
+export function logAbacusTsError(source: string, message: string, ...args: any[]) {
+  log(`${source}: ${message}`, undefined, { context: args });
+}
