+++ src/abacus-ts/rest/fills.ts
@@ -6,9 +6,9 @@ import { setAccountFillsRaw } from '@/state/raw';
 import { isTruthy } from '@/lib/isTruthy';
 
 import { refreshIndexerQueryOnAccountSocketRefresh } from '../accountRefreshSignal';
-import { loadableIdle } from '../loadable';
+import { loadableIdle } from '../lib/loadable';
 import { selectParentSubaccountInfo } from '../socketSelectors';
-import { createIndexerQueryStoreEffect } from './indexerQueryStoreEffect';
+import { createIndexerQueryStoreEffect } from './lib/indexerQueryStoreEffect';
 
 export function setUpFillsQuery(store: RootStore) {
   const cleanupListener = refreshIndexerQueryOnAccountSocketRefresh(['account', 'fills']);
