+++ src/abacus-ts/rest/orders.ts
@@ -7,9 +7,9 @@ import { setAccountOrdersRaw } from '@/state/raw';
 import { isTruthy } from '@/lib/isTruthy';
 
 import { refreshIndexerQueryOnAccountSocketRefresh } from '../accountRefreshSignal';
-import { loadableIdle } from '../loadable';
+import { loadableIdle } from '../lib/loadable';
 import { selectParentSubaccountInfo } from '../socketSelectors';
-import { createIndexerQueryStoreEffect } from './indexerQueryStoreEffect';
+import { createIndexerQueryStoreEffect } from './lib/indexerQueryStoreEffect';
 
 export function setUpOrdersQuery(store: RootStore) {
   const cleanupListener = refreshIndexerQueryOnAccountSocketRefresh(['account', 'orders']);
