+++ src/lib/telemetry.ts
@@ -4,7 +4,7 @@ import { isDev } from '@/constants/networks';
 import { track } from './analytics/analytics';
 import { dd } from './analytics/datadog';
 
-export const log = (location: string, error: Error, metadata?: object) => {
+export const log = (location: string, error?: Error, metadata?: object) => {
   if (isDev) {
     // eslint-disable-next-line no-console
     console.warn('telemetry/log:', { location, error, metadata });
